backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
pandas>=2.0
reportlab>=4.0
